relicensed WAL main definitions
