+++ LICENSING.md
@@ -10,6 +10,7 @@ The following folders and their subfolders are licensed under Apache-2.0:
 
 ```
 clients/
+pkg/ingester/wal
 pkg/logproto/
 pkg/loghttp/
 pkg/logqlmodel/
