+++ pkg/ingester/flush_test.go
@@ -2,6 +2,7 @@ package ingester
 
 import (
 	"fmt"
+	"github.com/grafana/loki/pkg/ingester/wal"
 	"os"
 	"sort"
 	"sync"
@@ -65,9 +66,9 @@ func TestChunkFlushingShutdown(t *testing.T) {
 
 type fullWAL struct{}
 
-func (fullWAL) Log(_ *WALRecord) error { return &os.PathError{Err: syscall.ENOSPC} }
-func (fullWAL) Start()                 {}
-func (fullWAL) Stop() error            { return nil }
+func (fullWAL) Log(_ *wal.WALRecord) error { return &os.PathError{Err: syscall.ENOSPC} }
+func (fullWAL) Start()                     {}
+func (fullWAL) Stop() error                { return nil }
 
 func Benchmark_FlushLoop(b *testing.B) {
 	var (
