+++ pkg/ingester/recovery.go
@@ -1,6 +1,7 @@
 package ingester
 
 import (
+	"github.com/grafana/loki/pkg/ingester/wal"
 	io "io"
 	"runtime"
 	"sync"
@@ -52,7 +53,7 @@ type Recoverer interface {
 	NumWorkers() int
 	Series(series *Series) error
 	SetStream(userID string, series record.RefSeries) error
-	Push(userID string, entries RefEntries) error
+	Push(userID string, entries wal.RefEntries) error
 	Done() <-chan struct{}
 }
 
@@ -152,7 +153,7 @@ func (r *ingesterRecoverer) SetStream(userID string, series record.RefSeries) er
 	return nil
 }
 
-func (r *ingesterRecoverer) Push(userID string, entries RefEntries) error {
+func (r *ingesterRecoverer) Push(userID string, entries wal.RefEntries) error {
 	return r.ing.replayController.WithBackPressure(func() error {
 		out, ok := r.users.Load(userID)
 		if !ok {
@@ -232,7 +233,7 @@ func (r *ingesterRecoverer) Done() <-chan struct{} {
 func RecoverWAL(reader WALReader, recoverer Recoverer) error {
 	dispatch := func(recoverer Recoverer, b []byte, inputs []chan recoveryInput) error {
 		rec := recordPool.GetRecord()
-		if err := decodeWALRecord(b, rec); err != nil {
+		if err := wal.DecodeWALRecord(b, rec); err != nil {
 			return err
 		}
 
@@ -267,7 +268,7 @@ func RecoverWAL(reader WALReader, recoverer Recoverer) error {
 				if !ok {
 					return
 				}
-				entries, ok := next.data.(RefEntries)
+				entries, ok := next.data.(wal.RefEntries)
 				var err error
 				if !ok {
 					err = errors.Errorf("unexpected type (%T) when recovering WAL, expecting (%T)", next.data, entries)
