+++ pkg/ingester/wal.go
@@ -2,6 +2,7 @@ package ingester
 
 import (
 	"flag"
+	"github.com/grafana/loki/pkg/ingester/wal"
 	"sync"
 	"time"
 
@@ -10,14 +11,13 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/prometheus/tsdb/wlog"
 
-	"github.com/grafana/loki/pkg/logproto"
 	"github.com/grafana/loki/pkg/util/flagext"
 	util_log "github.com/grafana/loki/pkg/util/log"
 )
 
 var (
 	// shared pool for WALRecords and []logproto.Entries
-	recordPool = newRecordPool()
+	recordPool = wal.NewRecordPool()
 )
 
 const walSegmentSize = wlog.DefaultSegmentSize * 4
@@ -54,16 +54,16 @@ func (cfg *WALConfig) RegisterFlags(f *flag.FlagSet) {
 type WAL interface {
 	Start()
 	// Log marshalls the records and writes it into the WAL.
-	Log(*WALRecord) error
+	Log(*wal.WALRecord) error
 	// Stop stops all the WAL operations.
 	Stop() error
 }
 
 type noopWAL struct{}
 
-func (noopWAL) Start()               {}
-func (noopWAL) Log(*WALRecord) error { return nil }
-func (noopWAL) Stop() error          { return nil }
+func (noopWAL) Start()                   {}
+func (noopWAL) Log(*wal.WALRecord) error { return nil }
+func (noopWAL) Stop() error              { return nil }
 
 type walWrapper struct {
 	cfg        WALConfig
@@ -102,7 +102,7 @@ func (w *walWrapper) Start() {
 	go w.run()
 }
 
-func (w *walWrapper) Log(record *WALRecord) error {
+func (w *walWrapper) Log(record *wal.WALRecord) error {
 	if record == nil || (len(record.Series) == 0 && len(record.RefEntries) == 0) {
 		return nil
 	}
@@ -117,7 +117,7 @@ func (w *walWrapper) Log(record *WALRecord) error {
 
 		// Always write series then entries.
 		if len(record.Series) > 0 {
-			buf = record.encodeSeries(buf)
+			buf = record.EncodeSeries(buf)
 			if err := w.wal.Log(buf); err != nil {
 				return err
 			}
@@ -126,7 +126,7 @@ func (w *walWrapper) Log(record *WALRecord) error {
 			buf = buf[:0]
 		}
 		if len(record.RefEntries) > 0 {
-			buf = record.encodeEntries(CurrentEntriesRec, buf)
+			buf = record.EncodeEntries(wal.CurrentEntriesRec, buf)
 			if err := w.wal.Log(buf); err != nil {
 				return err
 			}
@@ -165,56 +165,4 @@ func (w *walWrapper) run() {
 	)
 	checkpointer.Run()
 
-}
-
-type resettingPool struct {
-	rPool *sync.Pool // records
-	ePool *sync.Pool // entries
-	bPool *sync.Pool // bytes
-}
-
-func (p *resettingPool) GetRecord() *WALRecord {
-	rec := p.rPool.Get().(*WALRecord)
-	rec.Reset()
-	return rec
-}
-
-func (p *resettingPool) PutRecord(r *WALRecord) {
-	p.rPool.Put(r)
-}
-
-func (p *resettingPool) GetEntries() []logproto.Entry {
-	return p.ePool.Get().([]logproto.Entry)
-}
-
-func (p *resettingPool) PutEntries(es []logproto.Entry) {
-	p.ePool.Put(es[:0]) // nolint:staticcheck
-}
-
-func (p *resettingPool) GetBytes() []byte {
-	return p.bPool.Get().([]byte)
-}
-
-func (p *resettingPool) PutBytes(b []byte) {
-	p.bPool.Put(b[:0]) // nolint:staticcheck
-}
-
-func newRecordPool() *resettingPool {
-	return &resettingPool{
-		rPool: &sync.Pool{
-			New: func() interface{} {
-				return &WALRecord{}
-			},
-		},
-		ePool: &sync.Pool{
-			New: func() interface{} {
-				return make([]logproto.Entry, 0, 512)
-			},
-		},
-		bPool: &sync.Pool{
-			New: func() interface{} {
-				return make([]byte, 0, 1<<10) // 1kb
-			},
-		},
-	}
 }
