+++ pkg/ingester/wal/encoding.go
@@ -1,4 +1,4 @@
-package ingester
+package wal
 
 import (
 	"time"
@@ -9,6 +9,11 @@ import (
 
 	"github.com/grafana/loki/pkg/logproto"
 	"github.com/grafana/loki/pkg/util/encoding"
+)
+
+var (
+	// shared pool for WALRecords and []logproto.Entries
+	recordPool = NewRecordPool()
 )
 
 // RecordType represents the type of the WAL/Checkpoint record.
@@ -81,7 +86,7 @@ type RefEntries struct {
 	Entries []logproto.Entry
 }
 
-func (r *WALRecord) encodeSeries(b []byte) []byte {
+func (r *WALRecord) EncodeSeries(b []byte) []byte {
 	buf := encoding.EncWith(b)
 	buf.PutByte(byte(WALRecordSeries))
 	buf.PutUvarintStr(r.UserID)
@@ -95,7 +100,7 @@ func (r *WALRecord) encodeSeries(b []byte) []byte {
 	return encoded
 }
 
-func (r *WALRecord) encodeEntries(version RecordType, b []byte) []byte {
+func (r *WALRecord) EncodeEntries(version RecordType, b []byte) []byte {
 	buf := encoding.EncWith(b)
 	buf.PutByte(byte(version))
 	buf.PutUvarintStr(r.UserID)
@@ -136,7 +141,7 @@ outer:
 	return buf.Get()
 }
 
-func decodeEntries(b []byte, version RecordType, rec *WALRecord) error {
+func DecodeEntries(b []byte, version RecordType, rec *WALRecord) error {
 	if len(b) == 0 {
 		return nil
 	}
@@ -184,7 +189,7 @@ func decodeEntries(b []byte, version RecordType, rec *WALRecord) error {
 	return nil
 }
 
-func decodeWALRecord(b []byte, walRec *WALRecord) (err error) {
+func DecodeWALRecord(b []byte, walRec *WALRecord) (err error) {
 	var (
 		userID  string
 		dec     record.Decoder
@@ -200,7 +205,7 @@ func decodeWALRecord(b []byte, walRec *WALRecord) (err error) {
 		rSeries, err = dec.Series(decbuf.B, walRec.Series)
 	case WALRecordEntriesV1, WALRecordEntriesV2:
 		userID = decbuf.UvarintStr()
-		err = decodeEntries(decbuf.B, t, walRec)
+		err = DecodeEntries(decbuf.B, t, walRec)
 	default:
 		return errors.New("unknown record type")
 	}
